+++ packages/models-library/src/models_library/api_schemas_clusters_keeper/clusters.py
