+++ services/clusters-keeper/src/simcore_service_clusters_keeper/utils/clusters.py
@@ -3,11 +3,11 @@ import datetime
 import functools
 from typing import Any, Final
 
-from models_library.clusters import NoAuthentication
-from models_library.rpc_schemas_clusters_keeper.clusters import (
+from models_library.api_schemas_clusters_keeper.clusters import (
     ClusterState,
     OnDemandCluster,
 )
+from models_library.clusters import NoAuthentication
 from models_library.users import UserID
 from models_library.wallets import WalletID
 from types_aiobotocore_ec2.literals import InstanceStateNameType
